==> 01 data prep.ipynb <==
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix, roc_auc_score

### 1. importing dataset
* removed observations where IsBadBuy is missing

df = pd.read_csv('kickcars-sample-200.csv')
df = df[df['IsBadBuy'].notnull()]
df.shape

### 2. removing columns
* dropped PurchDate and VehYear as we already have VehicleAge
* dropped TopThreeAmericanName as as we already have Make
* dropped WheelTypeID as we already have WheelType
* dropped VNZIP1 as we already have VNST
* dropped Model and SubModel as there are multiple values
* dropped Trim as there are multiple values
* dropped Auction and IsOnlineSale as there are single values
* dropped Color and BYRNO as these may be misleading
* dropped AUCGUART as it is same as PRIMEUNIT

df = df.drop(['PurchDate','VehYear','WheelTypeID','TopThreeAmericanName','VNZIP1'], axis=1)
df = df.drop(['Model','SubModel','Trim'], axis=1)
df = df.drop(['Auction','IsOnlineSale','Color','BYRNO','AUCGUART'], axis=1)
df.shape

### 3. creating columns - numerical
* ratio of acquisition cost paid for the vehicle at time of purchase is taken
* if ratio is less than or equal to zero then it is made missing
* missing value is also imputed
* all positively related

df['MMRAcquisitionAuctionAveragePrice'] = df['VehBCost'] / df['MMRAcquisitionAuctionAveragePrice']
df['MMRAcquisitionAuctionCleanPrice'] = df['VehBCost'] / df['MMRAcquisitionAuctionCleanPrice']
df['MMRAcquisitionRetailAveragePrice'] = df['VehBCost'] / df['MMRAcquisitionRetailAveragePrice']
df['MMRAcquisitonRetailCleanPrice'] = df['VehBCost'] / df['MMRAcquisitonRetailCleanPrice']
df['MMRCurrentAuctionAveragePrice'] = df['VehBCost'] / df['MMRCurrentAuctionAveragePrice']
df['MMRCurrentAuctionCleanPrice'] = df['VehBCost'] / df['MMRCurrentAuctionCleanPrice']
df['MMRCurrentRetailAveragePrice'] = df['VehBCost'] / df['MMRCurrentRetailAveragePrice']
df['MMRCurrentRetailCleanPrice'] = df['VehBCost'] / df['MMRCurrentRetailCleanPrice']
df['WarrantyCost'] = df['WarrantyCost'] / df['VehBCost']

df['MMRAcquisitionAuctionAveragePrice'] = np.where(df['MMRAcquisitionAuctionAveragePrice']==np.inf, np.nan, df['MMRAcquisitionAuctionAveragePrice'])
df['MMRAcquisitionAuctionCleanPrice'] = np.where(df['MMRAcquisitionAuctionCleanPrice']==np.inf, np.nan, df['MMRAcquisitionAuctionCleanPrice'])
df['MMRAcquisitionRetailAveragePrice'] = np.where(df['MMRAcquisitionRetailAveragePrice']==np.inf, np.nan, df['MMRAcquisitionRetailAveragePrice'])
df['MMRAcquisitonRetailCleanPrice'] = np.where(df['MMRAcquisitonRetailCleanPrice']==np.inf, np.nan, df['MMRAcquisitonRetailCleanPrice'])
df['MMRCurrentAuctionAveragePrice'] = np.where(df['MMRCurrentAuctionAveragePrice']==np.inf, np.nan, df['MMRCurrentAuctionAveragePrice'])
df['MMRCurrentAuctionCleanPrice'] = np.where(df['MMRCurrentAuctionCleanPrice']==np.inf, np.nan, df['MMRCurrentAuctionCleanPrice'])
df['MMRCurrentRetailAveragePrice'] = np.where(df['MMRCurrentRetailAveragePrice']==np.inf, np.nan, df['MMRCurrentRetailAveragePrice'])
df['MMRCurrentRetailCleanPrice'] = np.where(df['MMRCurrentRetailCleanPrice']==np.inf, np.nan, df['MMRCurrentRetailCleanPrice'])
df['WarrantyCost'] = np.where(df['WarrantyCost']==np.inf, np.nan, df['WarrantyCost'])

df['MMRAcquisitionAuctionAveragePrice'] = df['MMRAcquisitionAuctionAveragePrice'].fillna(df['MMRAcquisitionAuctionAveragePrice'].mean())
df['MMRAcquisitionAuctionCleanPrice'] = df['MMRAcquisitionAuctionCleanPrice'].fillna(df['MMRAcquisitionAuctionCleanPrice'].mean())
df['MMRAcquisitionRetailAveragePrice'] = df['MMRAcquisitionRetailAveragePrice'].fillna(df['MMRAcquisitionRetailAveragePrice'].mean())
df['MMRAcquisitonRetailCleanPrice'] = df['MMRAcquisitonRetailCleanPrice'].fillna(df['MMRAcquisitonRetailCleanPrice'].mean())
df['MMRCurrentAuctionAveragePrice'] = df['MMRCurrentAuctionAveragePrice'].fillna(df['MMRCurrentAuctionAveragePrice'].mean())
df['MMRCurrentAuctionCleanPrice'] = df['MMRCurrentAuctionCleanPrice'].fillna(df['MMRCurrentAuctionCleanPrice'].mean())
df['MMRCurrentRetailAveragePrice'] = df['MMRCurrentRetailAveragePrice'].fillna(df['MMRCurrentRetailAveragePrice'].mean())
df['MMRCurrentRetailCleanPrice'] = df['MMRCurrentRetailCleanPrice'].fillna(df['MMRCurrentRetailCleanPrice'].mean())
df['WarrantyCost'] = df['WarrantyCost'].fillna(df['WarrantyCost'].mean())

df['VehicleAge_decades'] = df['VehicleAge']/10
df['VehOdo_lakhs'] = df['VehOdo']/100000
df.shape

### 4. creating columns - categorical
* converting categorical to numerical - Make, Transmission, WheelType, Nationality, Size, PRIMEUNIT, AUCGUART, VNST
* all are positively related - this is based on the imputation

df['Make'] = np.where(df['Make'].isin(['FORD','PONTIAC','SATURN']), 1, 
                      np.where(df['Make'].isin(['DODGE']), -1, 0))
df['Transmission'] = np.where(df['Transmission']=='AUTO', 1, -1)
df['WheelType'] = np.where(df['WheelType'].isnull(), 1, -1)
df['Nationality'] = np.where(df['Nationality']=='TOP LINE ASIAN', 1, -1)
df['Size'] = np.where(df['Size'].isin(['MEDIUM','COMPACT','LARGE','VAN']), 0, 
                      np.where(df['Size'].isin(['MEDIUM SUV']), 1, -1))
df['PRIMEUNIT'] = np.where(df['PRIMEUNIT'].isnull(), 1, -1)
df['VNST'] = np.where(df['VNST']=='VA', 1, -1)
df.shape

### 5. exporting dataset
* there are 11 numerical variables
* there are 7 categorical variables
* there are 2 other variables - RefId and IsBadBuy	
* also looking at univariate

df = df.drop(['VehicleAge','VehOdo','VehBCost'], axis=1)
df.to_csv('01 data prep.csv', index=False)
df.shape

df.describe(include='all').T

df.drop('RefId', axis=1).corr()
