==> 06 data holdout.ipynb <==
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix, roc_auc_score

### 1. importing dataset
* separating out x and y
* adding monotonic constraints

s0 = pd.read_csv('02 data split 0.csv')
s1 = pd.read_csv('02 data split 1.csv')
s2 = pd.read_csv('02 data split 2.csv')
df = pd.concat([s0,s1,s2])
df.shape

### 2. creating dataset
* using 10 variables
* total combinaions 2^10

Make2 = []
WheelType2 = []
MMRAcquisitionRetailAveragePrice2 = []
MMRAcquisitonRetailCleanPrice2 = []
MMRCurrentRetailAveragePrice2 = []
MMRCurrentRetailCleanPrice2 = []
VNST2 = []
WarrantyCost2 = []
VehicleAge_decades2 = []
VehOdo_lakhs2 = []

for Make1 in [1, -1]:
    for WheelType1 in [1, -1]:
        for MMRAcquisitionRetailAveragePrice1 in [0.664, 0.856]:
            for MMRAcquisitonRetailCleanPrice1 in [0.581, 0.726]:
                for MMRCurrentRetailAveragePrice1 in [0.681, 0.843]:
                    for MMRCurrentRetailCleanPrice1 in [0.594, 0.712]:
                        for VNST1 in [1, -1]:
                            for WarrantyCost1 in [0.126, 0.263]:
                                for VehicleAge_decades1 in [0.4, 0.6]:
                                    for VehOdo_lakhs1 in [0.611, 0.807]:
                                        
                                        Make2.append(Make1)
                                        WheelType2.append(WheelType1)
                                        MMRAcquisitionRetailAveragePrice2.append(MMRAcquisitionRetailAveragePrice1)
                                        MMRAcquisitonRetailCleanPrice2.append(MMRAcquisitonRetailCleanPrice1)
                                        MMRCurrentRetailAveragePrice2.append(MMRCurrentRetailAveragePrice1)
                                        MMRCurrentRetailCleanPrice2.append(MMRCurrentRetailCleanPrice1)
                                        VNST2.append(VNST1)
                                        WarrantyCost2.append(WarrantyCost1)
                                        VehicleAge_decades2.append(VehicleAge_decades1)
                                        VehOdo_lakhs2.append(VehOdo_lakhs1)

df = pd.DataFrame({'Make':Make2, 'WheelType':WheelType2, 'MMRAcquisitionRetailAveragePrice':MMRAcquisitionRetailAveragePrice2,
                  'MMRAcquisitonRetailCleanPrice':MMRAcquisitonRetailCleanPrice2, 'MMRCurrentRetailAveragePrice':MMRCurrentRetailAveragePrice2,
                  'MMRCurrentRetailCleanPrice':MMRCurrentRetailCleanPrice2, 'VNST':VNST2, 'WarrantyCost':WarrantyCost2,
                  'VehicleAge_decades':VehicleAge_decades2, 'VehOdo_lakhs':VehOdo_lakhs2}).reset_index()
df.shape

### 3. exporting dataset
* every 21th observation

df0 = df[df['index']%21==5]
df0 = df0.drop('index', axis=1)
df1 = df[df['index']%21==14]
df1 = df1.drop('index', axis=1)

df = pd.concat([df0,df1])
df.to_csv('06 data holdout.csv', index=False)
df.shape
